--- paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction-xml/src/main/res/layout/title_color.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:background="@android:color/white"
-    android:layout_width="match_parent"
-    android:padding="16dp"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:textColor="@android:color/black"
-        android:text="@string/color" />
-
-</LinearLayout>
