--- paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction-xml/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="color">Color</string>
-</resources>
