+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction/build.gradle
