+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction/src/main/res/layout/title_color.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent"
+  android:background="@android:color/white"
+  android:orientation="vertical"
+  android:padding="16dp">
+
+  <TextView
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="@string/color"
+    android:textColor="@android:color/black"
+    android:textSize="20sp" />
+
+</LinearLayout>
