+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction/src/main/res/values-ar/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <string name="color">اللون</string>
+</resources>
