+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction/src/test/resources/locale_ar.png
