+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/layout-direction/src/test/resources/locale_default_rtl.png
