--- paparazzi/paparazzi-gradle-plugin/src/test/projects/locale-qualifier-xml/src/main/res/values-en-rGB/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="color">Colour</string>
-</resources>
