+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/locale-qualifier/build.gradle
