+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/locale-qualifier/src/main/res/values-en-rGB/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <string name="color">Colour</string>
+</resources>
