+++ paparazzi/paparazzi-gradle-plugin/src/test/projects/locale-qualifier/src/test/resources/locale_default.png
