+++ paparazzi/paparazzi/src/main/java/app/cash/paparazzi/internal/SessionParamsBuilder.kt
@@ -97,12 +97,13 @@ internal data class SessionParamsBuilder(
     )
     result.fontScale = deviceConfig.fontScale
 
-    val locale = deviceConfig.localeQualifier
+    val localeQualifier = folderConfiguration.localeQualifier
+    val layoutDirectionQualifier = folderConfiguration.layoutDirectionQualifier
     // https://cs.android.com/android-studio/platform/tools/adt/idea/+/mirror-goog-studio-main:android/src/com/android/tools/idea/rendering/RenderTask.java;l=645
-    if (LayoutDirection.RTL == deviceConfig.layoutDirection && !Bridge.isLocaleRtl(locale.tag)) {
+    if (LayoutDirection.RTL == layoutDirectionQualifier.value && !Bridge.isLocaleRtl(localeQualifier.tag)) {
       result.locale = "ur"
     } else {
-      result.locale = locale.tag
+      result.locale = localeQualifier.tag
     }
     result.setRtlSupport(true)
